Fix config for Github Actions:
- Remove `coveralls-ruby` gem (not needed, not helpful).
- Add `simplecov-lcov` gem.
- Update version of `simplecov` gem from 0.16.x to 0.21.x. 0.16.x was throwing an error.
- Configure test suite to generate single-file LCOV report.
- Pass `path-to-lcov` to Coveralls Github Action.

Explanation: We don’t need coveralls-ruby gem for this integration. It duplicates the upload action, fails for the lack of repo-token, and doesn’t help because the Coveralls Github Action is looking for a LCOV file. We just need to generate a LCOV file from `simplecov` and then let the Coveralls Github Action covert & upload it to Coveralls.